Major revision to the microswitch carrier, I'm calling it version 2. The original version is still in the repository for those who might prefer it.
The carrier now prints in two pieces, and you have the option of either soldering wires directly to the switch as before, or using a 3 pin JST-XH plug in the back of the carrier.
It turns out that the spacing of pins 1 and 3 in the JST-XH exactly line up with the "common" and "normally-open" pins on the switch, and the carrier has the correct shape for the keyed connector to plug in.